# backend/Balenthiran.Snipit.WebApi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Balenthiran.Snipit.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/status": {
      "get": {
        "tags": [
          "StatusRoutes"
        ],
        "operationId": "GetStatus",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/transcriptions": {
      "post": {
        "tags": [
          "TranscriptionRoutes"
        ],
        "operationId": "SubmitTranscription",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionJob"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewLimitReached"
                }
              }
            }
          }
        }
      }
    },
    "/api/transcriptions/{id}": {
      "get": {
        "tags": [
          "TranscriptionRoutes"
        ],
        "operationId": "GetTranscriptionJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionJob"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/transcriptions/{id}/transcript": {
      "get": {
        "tags": [
          "TranscriptionRoutes"
        ],
        "operationId": "GetTranscript",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transcript"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/cuts": {
      "post": {
        "tags": [
          "CutRoutes"
        ],
        "operationId": "SubmitCut",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file",
                  "request"
                ],
                "type": "object",
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "file": {
                        "$ref": "#/components/schemas/IFormFile"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "request": {
                        "type": "string"
                      }
                    }
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CutJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/cuts/{id}": {
      "get": {
        "tags": [
          "CutRoutes"
        ],
        "operationId": "GetCutJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CutJobResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/cuts/{id}/download": {
      "get": {
        "tags": [
          "CutRoutes"
        ],
        "operationId": "DownloadCut",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CutJobResponse": {
        "required": [
          "downloadUrl",
          "id",
          "status",
          "error",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "downloadUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
      "JobStatus": {
        "enum": [
          "Pending",
          "Processing",
          "Completed",
          "Failed"
        ]
      },
      "PreviewLimitReached": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "SystemStatusResponse": {
        "required": [
          "version",
          "friendlyStatus",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "string"
          },
          "friendlyStatus": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Transcript": {
        "required": [
          "transcriptionJobId",
          "durationSeconds",
          "segments",
          "words"
        ],
        "type": "object",
        "properties": {
          "transcriptionJobId": {
            "type": "string",
            "format": "uuid"
          },
          "durationSeconds": {
            "type": "number",
            "format": "double"
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranscriptSegment"
            }
          },
          "words": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranscriptWord"
            }
          }
        }
      },
      "TranscriptionJob": {
        "required": [
          "id",
          "status",
          "error",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TranscriptSegment": {
        "required": [
          "index",
          "start",
          "end",
          "text"
        ],
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "format": "int32"
          },
          "start": {
            "type": "number",
            "format": "double"
          },
          "end": {
            "type": "number",
            "format": "double"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "TranscriptWord": {
        "required": [
          "text",
          "start",
          "end",
          "kept"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "start": {
            "type": "number",
            "format": "double"
          },
          "end": {
            "type": "number",
            "format": "double"
          },
          "kept": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "StatusRoutes"
    },
    {
      "name": "TranscriptionRoutes"
    },
    {
      "name": "CutRoutes"
    }
  ]
}